Apply AuthOperationFilter conditionally
All API endpoints required authentication until introducing the
`/api/operations` resource, which allows anonymous. This commit updates the
`IOperationFilter` to apply conditionally (when an `AuthorizeAttribute` is
present). It will result in the Swagger docs not marking the operation
endpoints with a lock/as behind authentication.
